Update remote_connection.py
Suggesting this change as when we try to run scripts on remote platform like browserstack and sauce labs, the urllib3 is throwing an error urllib3.exceptions.MaxRetryError: HTTPConnectionPool, this can be avoided if we set this value. the error was not coming until the urllib3==1.24.3. But post that urllib3 made certificate default.  There have been multiple communication chains regarding this issue. I could only fix this issue by explicity setting this value to none in the library stack.   In case you feel like not approving the change please do share the suggested fix to handle this issue as i could not find any.