[grid] Making EventBus more stable by ignoring events
Before, the EventBus was pulling and processing all events,
no matter if the component was meant to listen to them or not.
This caused issues when some messages were sent and not followed
the expected structure.

With this change, the component tells the bus to only process
events it listens to.

Helps with #10485