[grid] Closing input pipeline when client disconnects
This is needed when a client sends a request, and we
reply right away (e.g. session not found), but we do
not read the whole request content.

If we do not read the whole request content, Netty
will wait until it does, and then it blocks. Which
is not ideal since no more requests can be processed.

Now, we close the pipeline when the client disconnects.

Fixes #10485
Fixes https://github.com/SeleniumHQ/docker-selenium/issues/1439