An expectation for checking if non-empty text is present in element.
text_to_be_present_in_element_non_empty method is added to check that
the web element contains any non-empty text. Usually it's needed when
text is generated on-the-fly on the server side, and of course, that
the text of element is not empty.

Also, added self-test for this feature.