[grid] Printing readable exceptions
When triaging Grid issues, it is common
to see logs with exceptions generated in
the SeleniumSpanExporter. However, they
are json strings printed in a single line,
which makes them really hard to read.

With this change, if the event has information
about an exception, it will be printed in a
human-readable way.