Add docker device mapping configuration
Sometimes is needed to map a device file to a docker
container to make possible some tasks like virtualize
a VM or something else. Because of this, I believe it
is desirable to have a way to configure which device
files should be available in containers.

In this commit, we can define which device files should
be mapped in containers through the standard selenium grid
configuration file.