[iedriver] WM_User should be processed while waiting for a new window to appear.
A part of the fixes for issues that occur in Edge IE mode.

There is a thread that uses SendMessage to send WM_USER in the mouse click process, and this thread goes into a wait state.
If this is not done, the new browser window will never be displayed.
Therefore, you need to process the WM_USER message while waiting for a new window to appear.