[java] Removing deprecated Actions implementations
Deprecated in https://github.com/SeleniumHQ/selenium/commit/6b8bdae1f5ce8643571f519d3597f7f032732fd1
(1/7/2017) for selenium-3.1.0.

Some deprecated interfaces were not deleted but moved to the server package
because they are used by the RC package. And even if we are not
releasing RC anymore, we still want to keep the code for a bit longer
and then delete it all.