correctly checks the driver supports DevTools
When using a driver like EventFiringWebDriver as a parameter for JavaScriptEngine and trying to use StartEventMonitoring the constructor will throw a NullReferenceException. This is caused by casting the driver to IDevTools using "as" which will return null if the casting failed, and incorrectly checking the field "session" instead of "devToolsDriver" for null value.

fixes #10776