Add runtime dependency for selenium-webdriver
Selenium-devtools is an optional add-on for selenium-webdriver, thus the
only way that we can define interoperability is from a dependency on
webdriver from devtools.

In order for devtools to operate correctly with webdriver it needs to be
using a minimum version of 4.2 (as there was a dependency change of
websocket [1]).

As per the suggestion in
https://github.com/SeleniumHQ/selenium/pull/10868#issuecomment-1182518208
this adopts a pessimistic version constraint that will need to be
updated if selenium-webdriver has a major version update.

[1]: https://github.com/SeleniumHQ/selenium/commit/cd6989892592d99190010ccd9552858804d92d40