[grid] Default to Node healthcheck instead of initial status endpoint check
This will allow a Relay Node to boot and check periodically
for the status endpoint, and wait indefinitely for it to come
up.

If the Relay Node reports itself as DOWN because the endpoint
is not responding with 200, then the Node will eventually be
removed by the Distributor. However, when the endpoint responds
with 200 again, the heartbeat event will have UP as a status and
the Node will be registered again.

Fixes #10908