[grid] Forwarding requests with non-W3C locators
For context:
https://github.com/w3c/webdriver/issues/1649
https://github.com/appium/appium/issues/16550

Grid was only forwarding requests if they had
a W3C WebDriver locator strategy or if a custom
locator was matched. This was breaking Appium
because many locators over there are not W3C
WebDriver compliant. This change will let
determine the remote end if the locator is
valid or not.

Grid is still W3C WebDriver compliant with
this change because it is an intermediary
node https://w3c.github.io/webdriver/#dfn-intermediary-nodes.