Tighten visibility of rules in the support package
Because we create maven jars, we want to ensure that access
to targets is done via the `java_export` targets rather than
by referencing things directly. If we don't do this, there
is a risk that we'll include the same class in more than
one maven jar, leading to a "split package" and chaos.