Don't expect these attributes to magically chage their own type.
I think maybe this Enum setup could have been better served with some
constants and maybe a Literal, but I'm trying to trust what is here
already.

I am worried that there may be client code depending on being able to
hand strings in here, in which case almost everything about the
annotations on VirtualAuthenticatorOptions is wrong from the start.