[grid] Temporarily avoiding an extra jar merging
We are having issues while adding extra build info
into the MANIFEST.MF file while building the final
jar for Grid. This commit avoids the extra jar
merging that is producing this issue, and we will
likely revert this after we figure out how to solve
it. This was done to unblock the 4.5.1 release.

The MANIFEST.MF will be missing the git revision
and the date, but that information will be present
in the selenium-build.properties file anyway.