[grid] Allow users to wire-in custom routes
Grid3 allowed users to wire in custom servlets
So that a user could add customisations to the node.

Adding similar support to Grid4 via the interface
`org.openqa.selenium.grid.UserDefinedRoute` by
Leveraging Service Provider Interface driven wire-in
mechanism.

All the custom routes added by the user would be
available under the base path “/admin” in the Grid
eco-system depending on to which component the route
was added.