Allow specifying timeouts
Partially Fixes #11158

The client side read and connection timeouts are 
Currently hard coded and there’s no way to change them.

While we are working on creating a proper mechanism
To be able to specify timeouts when working with local
Drivers, it still will not allow a user to alter the 
Read and connection timeouts at the Node end when 
Running tests in Grid.

Exposing these two timeouts via JVM arguments so 
that a user may be able to set them when spawning
the node.

* `webdriver.client.connection.timeout.seconds` - for connection
timeouts in seconds
* `webdriver.client.read.timeout.minutes` - for read timeouts 
in minutes