[grid] Enabling `detectDrivers` original behaviour.
When Selenium Manager was integrated to the Java bindings,
Grid implicitly started using it, saving users time because
they did not need to configure drivers on the Node host
machine anymore.

This seemed ideal but made things harder for advanced users
who were configuring through the different flags for
drivers. For example, a user with a Windows machine, may
not want to use `IEDriverServer`.

This change introduces a new boolean flag `--selenium-manager`
set to `true` by default. When enabled, Selenium Manager will
be used to configure all the possible drivers on the host
machine. When disabled, the Node startup will only look for
drivers that are already present on the host machine.

This is only being implemented in Java because Grid has
that special use case. When using the bindings, users
will continue to pass the driver location or leave it on
the `PATH`to avoid Selenium Manager.

Fixes #11697