Replace GetCallingAssembly with GetExecutingAssembly
`GetCallingAssembly` is not supported with .NET 7 PublishAot.

It's not clear why `GetCallingAssembly` was chosen - `GetCallingAssembly` returns the assembly of the method that is calling this method. If two different assemblies were to call `ResourceUtilities.AssemblyVersion`, this code would cache the assembly of the first caller and produce what is essentially an incorrect result for the second caller.

Given the only caller of `ResourceUtilities.AssemblyVersion` is within the same assembly, it's likely `Assembly.GetExecutedAssembly()` was intended (it's also indicated by the name of the local variable).

Fixes dotnet/runtime#86202.
Fixes #8549 (that was won't fixed but it's the same problem).