[dotnet] Simplfy to only target .NET Standard 2.0
Rather than targeting multiple individual target framework versions,
simplify the project targeting to only target .NET Standard 2.0. Doins
so, the resulting assemblies will be usable by projects targeting any of
the following versions:

* .NET Framework: 4.6.1, 4.6.2, 4.7, 4.7.1, 4.7.2, 4.8, 4.8.1
* .NET Core: 2.0, 2.1, 2.2, 3.0, 3.1
* .NET versions: 5.0, 6.0, 7.0

Additionally, they will be usable from the Mono and Xamarin frameworks.

This will simplify the build process and maintenance going forward, and
will allow the project to greatly reduce the complexity of the Bazel
builds.

As a note, this does remove support for .NET 4.5.x, the last version
of which (4.5.2) was released in 2014, and reached end-of-life in
April 2022.