Add assignable Session ID to Webdriver
A Webdriver normally generates a new Session Id everytime it is a new instance.
In a case where you want to do multiple tests, with selenium grid, if you have a stream
of automation steps that wish to continue on the same Driver/Session, you would have to cache
an instance of that webdriver locally and wait for the next call. Once a continuation call
has been made, you would have to fetch the instance of the webdriver,
with its Session Id intact. This is obviously not quick and also not a good practice
for memory as you would have to cache multiple driver
instances for multiple tests, running at the same time.

To avoid all of this, we have added an extra parameter to the constructor of the WebDriver.
The session Id property is then assigned the already existing session, so it can point to it.
This means that for every session that is created by the webdriver, we can
return the session Id or cache it so that if a future call wants to
continue on the same webdriver session on Selenium Grid, it can do so by passing
in the preferred session Id which will then point it to an already existing webdriver
and automation can continue.

An example of this would be when you have a website that needs to authenticate a user.
Automation Method 1: Ensures that the user has logged in.
If that user has logged in, return a success to the automation API
with the webdriver session Id. Ensure that this process has been logged correctly.
Then call the continuation method to navigate on the site, which will be the next method.

Automation Method 2: Ensure that the user is able to redirect to the contact-us page,
as they are already logged in. Automate filling in the contact-us form and submit, and return
a success message message. Then trigger a driver.Quit() when the above steps are done.

As you can see, the above steps require the calls to be pointed to the same webdriver session
in order for the automation to be completed.