[java] Allow setting proxy for all http client instances (#12940)
he default AsyncHttpClient used in the previous version of Selenium allowed the use of a global proxy set via the system properties. https://github.com/SeleniumHQ/selenium/blob/selenium-4.13.0/java/src/org/openqa/selenium/remote/http/netty/NettyClient.java#L88C3-L88C3

Extending the same functionality for the current default JDK 11 Http Client.

However, we have the option to use client config to set proxy in the RemoteWebDriverBuilder. It is used by the underlying HttpClient of RemoteWebDriver only. The other HttpClients created along the flow, do not use it. Example:
https://github.com/SeleniumHQ/selenium/blob/trunk/java/src/org/openqa/selenium/devtools/CdpEndpointFinder.java#L48

Similarly, for the Grid, we create different instances of HttpClient to be used in different classes which cannot be configured to use the proxy.