"Get Text" atom has to return the visible text from an element within a closed ShadowDOM.
As discussed on issue #13132 the "Get Text" atom currently
fails for elements that are located within a closed ShadowDOM.
In those cases a check for ".shadowRoot" will not succeed
because it's "null". This patch fixes the check by making sure
that it is only run when a valid ShadowRoot is actually available.

 # Please enter the commit message for your changes. Lines starting