[rb] Fully support Chrome 120+ old headless mode
When Chrome's new headless mode was announced, it was mentioned that
they intended to remove the old headless mode from the Chrome binary.
Some people still want to use the old headless mode, as it's more
lightweight, performant, and has fewer dependencies - at the expense of
having fewer features.

Chrome 120+ has a standalone `chrome-headless-shell` binary to keep the
old headless mode working for those who wish to use it. If your Ruby
code sets up the Chrome driver with `--headless` (which currently
defaults to `old`) or `--headless=old` explicitly, you will lose
capabilities such as setting permissions, fetching the logs and using
the CDP API.

That happens because the new binary has a browser name of
`chrome-headless-shell`, instead of `chrome`. That makes it be
recognized as something other than Chrome, not setting its capabilities.

This commit introduces `chrome-headless-shell` as an alternative Chrome
browser name, fixing this issue.

Fixes #13112