Replace size checks with isEmpty method
Replaced all the instances where the size of a collection was being checked to see if it was greater than zero, with the isEmpty method. The isEmpty method is more efficient and performant because it doesn't need to count all elements in order to determine if the collection has any elements, it just checks if there is at least one.