[java] Refactoring to check for emptiness (#13335)
* Optimized checking for empty collections in ChromeDriverFunctionalTest

Replaced size-based checks for collection emptiness with isEmpty method calls in ChromeDriverFunctionalTest class. This change simplifies the code and potentially improves performance by avoiding unnecessary size calculations when just checking if the collection is empty or not.

* Replace size checks with isEmpty method

Replaced all the instances where the size of a collection was being checked to see if it was greater than zero, with the isEmpty method. The isEmpty method is more efficient and performant because it doesn't need to count all elements in order to determine if the collection has any elements, it just checks if there is at least one.