Use all provided options while merging them for Firefox (#13582)
* Use all provided options while merging them for Firefox

This fix addresses the discrepancy that we have in methods of processing browser options for Chromium and Firefox. In the case of Chromium, if the option name doesn't match any of previously processed, we add it anyway. It doesn't happen for Firefox. So in the case when one option is set in stereotype, but other is provided in capabilities this second one is missed from the result.

Unit test updated to cover this case.

* fixed formatting in SessionCapabilitiesMutatorTest