*Changes Done: - Migrate the String.Format to Formatter.Format in : - ChromeDriverSerivce.java - EdgeDriverService.java - RemoteWebDriver.java
*Issue Explaination:
- Having System Date & Language in Arabic caused with driver setup.
Using String.Format for ports and some other lines caused them to be localized in the system language and in arabic case it caused the numbers and some other character to be localized in arabic,
e.g. the port. And that caused an issues when appending the port to the URL in order to setup the browser

*Fix explaination:
- Now it's utilizaing the Formatter.Format and forcing the localization to be in English.US in the Formatter Constructor