Added step in api doc build to regenerate the autodoc stub pages
Instead of using the commited stub files which contain "pre-compiled"
outlines for the api doc this will regenrate them. It should be noted
this might add some time to the build process so a future change might
be to see how store and update either commited files or previous
build artifacts.

I also and working though the output directory which is currently set to
`-o docs/source`. This seem to dump all the sub files into the main source
directory insted of neatly organied into child directories. Going to try
to reowrk this but wanted to commit so as not to lose how I did this the
first time ;)