Updates the py api doc build (#14173)
* Added requirements.txt to doc build environment

For the API documentation as sphinx autogenerates the
api docs from the code it needs the dependent Python
packages. This change adds those to the tox docs build
environment.

* Updated Sphinx version within docs/requirements.txt

Bring Sphinx version up to the latest Python 3.8 suported version,
v7.1.2. This change was made within py/docs/requirements.txt. It
should be noted that there is a seperate version requirement given
within py/tox.ini which I need to test and see which one is grabed
using the selenium build process. (I supsect it will be the one
within py/tox.ini.)

* Added step in api doc build to regenerate the autodoc stub pages

Instead of using the commited stub files which contain "pre-compiled"
outlines for the api doc this will regenrate them. It should be noted
this might add some time to the build process so a future change might
be to see how store and update either commited files or previous
build artifacts.

I also and working though the output directory which is currently set to
`-o docs/source`. This seem to dump all the sub files into the main source
directory insted of neatly organied into child directories. Going to try
to reowrk this but wanted to commit so as not to lose how I did this the
first time ;)

* wip adding new files in api.rst

Credit goes to @iampopovich for these changes

* wip alphabetical order in api.rst

Credit goes to @iampopovich for these changes

* Removed previous generated api doc stub files

Prefer instead to auto-generate these as part of the build.
Also ignore those sub-directories.

* Updated the tox api doc build steps

- Use the docs/requirements.txt file instead of repeating the build
  versions for jinja and sphinx
- Removed the output directory option on sphinx-autogen as that was
  putting everything in a flat directory structure at the root instead
  of the well organized folders