[bazel] Use the `release` configs for the release tasks
Rather than passing `--stamp` everywhere, use a proper config
so that we can easily add additional flags.

I've broken out a `remote_release` config which builds things
remotely and then downloads the artifacts given on the command
line to the local machine. This only makes sense for targets
which we do _not_ `run` (as `run` happens locally, so we'd
need to rebuild any artifacts anyway)