chore(ci): correct github-script API calls (#14442)
* chore(ci): correct github-script API calls

Since V5 of github-script the Octokit context available via `github` no
longer has REST methods directly on it, they were moved to
`github.rest.*` instead. Update the references in delete-comments.yml
job to match.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>

* fix(ci): remove 'Download' from delete-comments

This is too generic a word and frequently matches against comments that
don't need to be deleted, nor should the user be blocked as the current
workflow will do.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>

* fix(ci): correct block user task

The existing code was calling the individual "block a user" REST
endpoint with incorrect parameters and never would have worked. Update
it to (presumably achieve the desired outcome) block the user from the
owning organisation instead.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>

---------

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>