[py] Ignore exception in no_driver_after_test fixture
The `no_driver_after_test` indicates that tests are expected
to quit the driver themselves, leaving `selenium_driver` with a stale
driver instance.

This commit keeps the `stop_driver()` call as a safety fallback, but
ignores WebDriverExceptions since they're expected in the call when
the test properly quits the driver.