Update dev-environment-setup.ps1
This PR fixes the parameters passed to the `choco` command. Because of the way powershell sends parameters to a command when it is external (https://github.com/MicrosoftDocs/PowerShell-Docs/issues/2361), the `params` parameters passed where not being correctly interpreted by `choco` and would therefore not being taken into account. Now, we use `Start-Process` with `-ArgumentList` which correctly parses and sends the parameters to `choco.exe` which allow us a couple of things:

1. `msys2` will be installed to the `InstallDir´ (it just so happens that we use the default folder for installation)
2. We can not add workloads to the VS 2022 Community installation, which has a couple of benefits:
2.1. We can add `Microsoft.VisualStudio.Workload.NativeDesktop;includeRecommended` and `Microsoft.VisualStudio.Workload.ManagedDesktop` which are necessary for compiling Selenium in Visual Studio
2.2. Now the installation of VS is completely automatic and does not need feedback from the user