[rb] Remove `prism` dependency
This project does not use prism at all. At most it is being used by rubocop during local development.

It was added in https://github.com/SeleniumHQ/selenium/pull/16357 but it is unclear to me why.
I just bumped it for this repo, perhaps that JRuby issue has since been fixed.
Either way, it should not be part of the gemspec.