Avoid `sleep` in drag'n'drop tests
fix the test page usability: make the elements visible only after they are fully initialized.

To force revealing potential flaky tests in the future, I've added a small pause for the initialization code. So the tests should wait until the element appears on the screen - then it's ready for manipulations.