[bazel+closure]: Vendor the version of closure library we use
We used to have a version of Google Closure Library tucked away in
`//third_party/closure/goog` We never really updated it, because
we switched to using `rules_closure`, and that bundles a different
version of Google Closure Library.

This PR vendors the version of Closure back into
`third_party/closure/goog` and then updates our builds and tests
to use.

This is a necessary prelude to updating `rules_closure` to the
latest release. That version no longer bundles a copy of closure,
so we need to be using a vendored one. However, the version we end
up on won't be _this_ version of closure because the later
`rules_closure` has an updated closure compiler that doesn't like
our code.

However, one thing at a time! Let's start by vendoring the thing
we're using and remove ourselves from the clutches of a bundled
version of the library.