Simplify creating lists in ByChainedTest
After the simplification, it became obvious that some of ByChainedTest tests don't really test anything. They create a bunch of lists (e.g. `elems345`) which are not used. And only the empty list of elements is loaded and asserted.

Seems that this complex/misleading test logic in ByChainedTest was
introduced in commit c4861b86 (some "patch from BenChambers"). Needs to be reviewed and fixed.